+++ src/plugin.cpp
@@ -38,4 +38,5 @@ void init(Plugin* p) {
 	p->addModel(modelCVrouter);
 	p->addModel(modelAdder8);
 	p->addModel(modelClocker);
+	p->addModel(modelSickoSampler2);
 }