+++ changelog.md
@@ -1,4 +1,9 @@
-### 2.5.5 (2023-07)
+### 2.5.6 (2023-09-xx)
+- added sickoSampler2 module
+- fixed a bug on sickoSampler crashing rack when pressing nav button after a new recording
+
+
+### 2.5.5 (2023-07-17)
 - fixed a bug on all player/sampler modules crashing VCV on MACs when root folder is not found
 - added a functionality on all player/sampler modules to to show the path of not found files